pyproject.toml
[project]
name = "openai-agents-apps"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Muhammad", email = "hamzaadil56@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.4.1",
    "ipykernel>=6.29.5",
    "openai-agents>=0.0.4",
    "python-dotenv>=1.0.1",
]

[project.scripts]
openai_agents_apps = "openai_agents_apps:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
